# salon-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salon API",
        "description": "API za upravljanje frizerskim salonom",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/available-slots": {
            "get": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Dostupni termini frizera",
                "operationId": "5eaf84c5da590299e23852bb6319499a",
                "parameters": [
                    {
                        "name": "hairdresser_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2026-03-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista slobodnih termina"
                    },
                    "404": {
                        "description": "Frizer ne radi taj dan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservations": {
            "get": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Lista svih rezervacija",
                "operationId": "f10e68ff0ce263ef754b2f640eda8d93",
                "responses": {
                    "200": {
                        "description": "Lista rezervacija"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Kreiranje nove rezervacije",
                "operationId": "4f59911f11563f79b4f01bd02222607d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hairdresser_id",
                                    "service_id",
                                    "start_time"
                                ],
                                "properties": {
                                    "hairdresser_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "2026-03-10 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rezervacija kreirana"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Detalji rezervacije",
                "operationId": "14c077250033bb10a7305f9724f10bd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji rezervacije"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Ažuriranje statusa rezervacije",
                "operationId": "453c2db0a909455ff1a629a0ed80f9a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status ažuriran"
                    },
                    "403": {
                        "description": "Nemate dozvolu"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Brisanje rezervacije",
                "operationId": "3f3a1d6e7ffee88b7ef58128a7f9295c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rezervacija obrisana"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservations/{id}/cancel": {
            "put": {
                "tags": [
                    "Rezervacije"
                ],
                "summary": "Otkazivanje rezervacije",
                "operationId": "69570aef721eb139fc2176bf30ff90f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rezervacija otkazana"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}